move file manipulation function in utils
