+++ rest_api/utils.py
@@ -1,8 +1,25 @@
+import json
+import os
+import random
+import base64
 from users.models import User
 from users.serializers import UserSerializer
 from rest_api.serializers import DepartmentSerializer, CourseSerializer, FileSerializer
 from users.signals import notification_handler
 from rest_api.models import Course
+from studyportal.drive.drive import driveinit
+from studyportal.settings import CUR_DIR
+from apiclient.http import MediaFileUpload
+
+
+STRUCTURE = os.path.join(
+    CUR_DIR,
+    'drive/structure.json'
+)
+STRUCTURE_TEST = os.path.join(
+    CUR_DIR,
+    'test/resources/structure.json'
+)
 
 
 def add_course(course, department):
@@ -37,6 +54,65 @@ def add_file(file, course):
                                      verb="added a file", action=file_data['title'],
                                      notification_type="addfile", target=course,
                                      link="/departments/" + department_code + "/courses/" + file_data['course']['code'])
+
+
+def uploadToDrive(service, folder_id, file_details):
+    file_metadata = {
+        'name': file_details['name'],
+        'parents': [folder_id]
+    }
+    media = MediaFileUpload(
+        file_details['location'],
+        mimetype=file_details['mime_type']
+    )
+    file = service.files().create(
+        body=file_metadata,
+        media_body=media,
+        fields='id'
+    ).execute()
+    return file.get('id')
+
+
+def get_file_details(file, name, filetype, course, for_review):
+    # For local dev change 'STRUCTURE' to 'STRUCTURE_TEST'
+    with open(STRUCTURE_TEST) as f:
+        structure = json.load(f)
+    # File manipulation starts here
+    file_type = file.split(",")[0]
+    mime_type = file_type.split(":")[1].split(";")[0]
+    ext = file_type.split("/")[1].split(";")[0]
+    base64String = file.split(",")[1]
+
+    rand = str(random.randint(0, 100000))
+    temp = open("temp" + rand + "." + ext, "wb")
+    temp.write(base64.b64decode(base64String))
+    file_details = {
+        'name': name,
+        'mime_type': mime_type,
+        'location': "temp" + rand + "." + ext
+    }
+    file_size = os.path.getsize("temp" + rand + "." + ext)
+    size = get_size(file_size)
+
+    # Get folder id from config
+    if for_review:
+        review_identifier = str("_review")
+    else:
+        review_identifier = str("")
+    folder_identifier = filetype.lower().replace(" ", "") + review_identifier
+    folder_id = structure['study'][course.department.abbreviation][course.code][folder_identifier]
+    driveid = uploadToDrive(
+        driveinit(),
+        folder_id,
+        file_details
+    )
+    os.remove("temp" + rand + "." + ext)
+    # end of manipulation
+    return {
+        'size': size,
+        'driveid': driveid,
+        'ext': ext
+    }
 
 
 def get_size(size):
